Cositas menores

+++ routes/web.php
@@ -26,9 +26,9 @@ Route::get('usuarios', [App\Http\Controllers\UserController::class, 'index']) ->
 
 Route::get('usuarios/create', [App\Http\Controllers\UserController::class, 'create'])->name('usuarios');
 Route::post('usuarios',[UserController::class, 'store'])->name('usuarios.store');
-Route::get('usuarios/{id}/edit',[UserController::class, 'edit'])->name('usuarios.edit');
-Route::put('usuarios/{id}',[UserController::class, 'update'])->name('usuarios.update');
-Route::delete('usuarios/{id}',[UserController::class, 'destroy'])->name('usuarios.destroy');
+Route::get('usuarios/{id}/edit',[UserController::class, 'edit'])->name('usuarios.edit')->middleware('auth');
+Route::put('usuarios/{id}',[UserController::class, 'update'])->name('usuarios.update')->middleware('auth');
+Route::delete('usuarios/{id}',[UserController::class, 'destroy'])->name('usuarios.destroy')->middleware('auth');
 Route::get('usuarios/{id}/show',[UserController::class,'show'])->name('usuarios.show');
 
 Route::get('/dino', function () {
